+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [1.6.2]
 - Skip `WeatherConverter` if failed
 - Fixes buggy zip output by using temp folder again like before v1.6.0
+- Keep old `assets` folder
 
 ## [1.6.1]
 - Fix