+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## [1.6.3]
+- Fix `publish` script
+
 ## [1.6.2]
 - Skip `WeatherConverter` if failed
 - Fixes buggy zip output by using temp folder again like before v1.6.0